Removed unnecessary escape characters
There are some backslashes in there that are rendered to the html as well. Adding the backslash to the filter config will cause the field to not be created.

I haven't tested the resulting Ruby code but I can see that in other files the underscores are not escaped.

Fixes #9